Fix MSVC Analyzer overflow warning

+++ utils.c
@@ -508,7 +508,7 @@ uint64_t microseconds(void) {
 	if (gettimeofday(&tv, NULL) < 0) {
 		return 0;
 	} else {
-		return tv.tv_sec * 1000000 + tv.tv_usec;
+		return (uint64_t)tv.tv_sec * 1000000 + tv.tv_usec;
 	}
 #endif
 }
